+++ apps/mobile-wallet/src/features/splash-screen/SplashScreen.tsx
@@ -16,12 +16,26 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-export type Language = 'en-US' | 'fr-FR' | 'el-GR' | 'vi-VN' | 'pt-PT'
+import { Image } from 'expo-image'
+import styled from 'styled-components/native'
 
-export const languageOptions: { label: string; value: Language }[] = [
-  { label: 'English', value: 'en-US' },
-  { label: 'Français', value: 'fr-FR' },
-  { label: 'Ελληνικά', value: 'el-GR' },
-  { label: 'Tiếng Việt', value: 'vi-VN' },
-  { label: 'Português', value: 'pt-PT' }
-]
+import splashSrc from '~/features/splash-screen/splash.png'
+
+const SplashScreen = () => (
+  <SplashScreenStyled>
+    <SplashScreenImage source={splashSrc} contentFit="center" style={{ objectFit: 'contain' }} />
+  </SplashScreenStyled>
+)
+
+export default SplashScreen
+
+const SplashScreenStyled = styled.View`
+  flex: 1;
+  justify-content: center;
+  align-items: center;
+`
+
+const SplashScreenImage = styled(Image)`
+  flex: 1;
+  width: 100%;
+`