+++ apps/mobile-wallet/src/screens/Dashboard/DashboardScreen.tsx
@@ -39,6 +39,7 @@ import WalletSwitchButton from '~/components/WalletSwitchButton'
 import BuyModal from '~/features/buy/BuyModal'
 import FundPasswordReminderModal from '~/features/fund-password/FundPasswordReminderModal'
 import { useAppSelector } from '~/hooks/redux'
+import { useAsyncData } from '~/hooks/useAsyncData'
 import { InWalletTabsParamList } from '~/navigation/InWalletNavigation'
 import { ReceiveNavigationParamList } from '~/navigation/ReceiveNavigation'
 import { SendNavigationParamList } from '~/navigation/SendNavigation'
@@ -68,7 +69,7 @@ const DashboardScreen = ({ navigation, ...props }: ScreenProps) => {
   const [isBackupReminderModalOpen, setIsBackupReminderModalOpen] = useState(!isMnemonicBackedUp)
   const [isSwitchNetworkModalOpen, setIsSwitchNetworkModalOpen] = useState(false)
   const [isBuyModalOpen, setIsBuyModalOpen] = useState(false)
-  const [isNewWallet, setIsNewWallet] = useState(false)
+  const { data: isNewWallet } = useAsyncData(getIsNewWallet)
 
   const buttonsRowStyle = useAnimatedStyle(() => ({
     height: withDelay(800, withSpring(65, defaultSpringConfiguration)),
@@ -82,17 +83,14 @@ const DashboardScreen = ({ navigation, ...props }: ScreenProps) => {
   }, [isMnemonicBackedUp, needsFundPasswordReminder])
 
   useEffect(() => {
+    if (!isNewWallet) return
+
     try {
-      getIsNewWallet().then((isNew) => {
-        if (isNew !== undefined) {
-          setIsNewWallet(isNew)
-          storeIsNewWallet(false)
-        }
-      })
+      storeIsNewWallet(false)
     } catch (e) {
       console.error(e)
     }
-  }, [])
+  }, [isNewWallet])
 
   const handleReceivePress = () => {
     if (addressHashes.length === 1) {